centaur

+++ ft_struct.h
@@ -6,7 +6,7 @@
 /*   By: sabrenda <sabrenda@student.21-school.ru    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/13 15:03:41 by sabrenda          #+#    #+#             */
-/*   Updated: 2021/01/16 07:29:21 by sabrenda         ###   ########.fr       */
+/*   Updated: 2021/01/16 22:51:46 by sabrenda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,8 +35,8 @@ typedef struct				s_flag
 	char					type_c;
 	char					*type_s;
 	unsigned long long int	type_p;
-	int						type_i;
-	unsigned int			type_u;
+	long long int			type_i;
+	unsigned long long int	type_u;
 	unsigned long long int	type_x;
 	unsigned int			type_n;
 	float					type_f;
